==> 06-featureextract-data.py <==
# Full: All 64 channels
# Moderate: F3, F1, Fz, F2, F4, FC3, FC1, FCz, FC2, FC4, C3, C1, Cz, C2, C4, CP3, CP1, CP2, and CP4 (Si-Mohammed et al., 2020)
# Minimal: Only FCz and Cz (Torres et al., 2018)

# Also normalize epochs for ML
